# docsandtests/eddata_apidoc.json
{
  "swagger": "2.0",
  "info": {
    "description": "EdData API is a secure open source api server that schools can use to create integrations between edtech vendors of their choosing, and also a common api for developers to write their applications against",
    "version": "1.0.0",
    "title": "EdData API",
    "termsOfService": "http:\/\/api.psd401.net\/terms\/",
    "contact": {
      "name": "PSD Technical Services Team",
      "url": "https://opentech.psd401.net",
      "email": "hagelk@psd401.net"
    },
    "license": {
      "name": "MIT",
      "url": "https://raw.githubusercontent.com/psd401/eddata-api/master/LICENSE"
    }
  },
  "host": "api2.psd401.net",
  "basePath": "\/api\/v1",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "\/computers": {
      "post": {
        "tags": [
          "computers"
        ],
        "summary": "Add a new computer to the database",
        "description": "",
        "operationId": "addComputer",
        "consumes": [
          "application\/json"
        ],
        "produces": [
          "application\/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Computer object that needs to be added to the database",
            "required": false,
            "schema": {
              "$ref": "#\/definitions\/Computer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#\/definitions\/Computer"
            }
          }
        }
      },
      "put": {
        "tags": [
          "computers"
        ],
        "summary": "Add a new computer or update existing computers in the database",
        "description": "Looks up computers by their snum(serialNumber) to do a comparison. Will update if they exist, otherwise will create new.",
        "operationId": "addUpdateComputer",
        "consumes": [
          "application\/json"
        ],
        "produces": [
          "application\/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Computer object that needs to be added or updated in the database",
            "required": false,
            "schema": {
              "$ref": "#\/definitions\/Computer"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Computer Not Found"
          },
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#\/definitions\/Computer"
            }
          }
        }
      },
      "get": {
        "tags": [
          "computers"
        ],
        "summary": "Finds All Computers",
        "description": "",
        "operationId": "findComputers",
        "produces": [
          "application\/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "description": "offset value for records returned",
            "required": false,
            "type": "integer"
          },
          {
            "in": "query",
            "name": "limit",
            "description": "limit value for records returned",
            "required": false,
            "type": "integer"
          },
          {
            "in": "query",
            "name": "ids",
            "description": "array of ids to find a specific set of records",
            "required": false,
            "type": "array",
            "items": {
              "type": "integer"
            },
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "snum",
            "description": "serial number search",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "user",
            "description": "last logged on user search",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "ip",
            "description": "ip address search",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#\/definitions\/Computer"
              }
            }
          },
          "404": {
            "description": "Computers not found"
          }
        }
      }
    },
    "\/computers\/{id}": {
      "get": {
        "tags": [
          "computers"
        ],
        "summary": "Find computer by ID",
        "description": "",
        "operationId": "getComputerById",
        "produces": [
          "application\/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ID of computer that needs to be fetched",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "404": {
            "description": "Computer not found"
          },
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#\/definitions\/Computer"
            }
          }
        }
      },
      "put": {
        "tags": [
          "computers"
        ],
        "summary": "Update an existing computer",
        "description": "",
        "operationId": "updateComputer",
        "consumes": [
          "application\/json"
        ],
        "produces": [
          "application\/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Computer object that needs to be updated in the database",
            "required": false,
            "schema": {
              "$ref": "#\/definitions\/Computer"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Computer not found"
          },
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#\/definitions\/Computer"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "computers"
        ],
        "summary": "Deletes a computer",
        "description": "",
        "operationId": "deleteComputer",
        "produces": [
          "application\/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Computer id to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ]
      }
    },
...
  },
...
}